Report validation metrics on invalidated network
Orignially Iwlan will only report metric when current network is validated. However, it will cause bias in the metrics and affect the result. Reporting the metrics regardless of the current validation status.

Add a boolean validationTriggered to indicate whether the result is from a new instance of validation or from previous.

Bug: 394400931
Flag: EXEMPT metrics changes
Test: atest IwlanTests, on device test
Change-Id: Ic6ab32915e97bb0e8d725fddb39a855a59e432e2